+++ lib/dart_foundation.dart
@@ -2,8 +2,18 @@ library;
 
 export 'src/errors/value_error.dart';
 export 'src/failures/value_failure.dart';
+
+// Types
 export 'src/types/value_object.dart';
-export 'src/validator/validator.dart';
-export 'src/value_objects/email_address.dart';
-export 'src/value_objects/password.dart';
+
+// ValueObjects
+export 'src/value_objects/timestamp.dart';
 export 'src/value_objects/unique_id.dart';
+
+// Validators
+export 'src/validators/validator.dart';
+export 'src/validators/bool_validator.dart';
+export 'src/validators/datetime_validator.dart';
+export 'src/validators/double_validator.dart';
+export 'src/validators/int_validator.dart';
+export 'src/validators/string_validator.dart';