+++ lib/src/validators/bool_validator.dart
@@ -10,7 +10,19 @@ class BoolValidator {
   /// List of rules to be validated.
   final List<BoolRule> _rules = [];
 
+  /// Public constructor for creation using [Validator] class.
   BoolValidator();
+
+  /// Bool validation method to check if the input is equal to another bool.
+  /// Adds specified rule to ruleset.
+  BoolValidator equals(final bool equals) {
+    _rules.add((input) {
+      return input == equals
+          ? right(input)
+          : left(ValueFailure('Bool is not equal to ${equals.toString()}'));
+    });
+    return this;
+  }
 
   /// Validate method that is taking the input value to validate.
   /// Every defined rule is applied to the input bool.