+++ lib/src/validators/validator.dart
@@ -1,13 +1,12 @@
 import 'package:dart_foundation/dart_foundation.dart';
-import 'package:dart_foundation/src/validator/bool_validator.dart';
-import 'package:dart_foundation/src/validator/int_validator.dart';
-import 'package:dart_foundation/src/validator/string_validator.dart';
 import 'package:dartz/dartz.dart';
 
 class Validator {
   Validator._();
 
   static BoolValidator bool() => BoolValidator();
+  static DatetimeValidator datetime() => DatetimeValidator();
+  static DoubleValidator double() => DoubleValidator();
   static IntValidator int() => IntValidator();
   static StringValidator string() => StringValidator();