+++ test/src/validator/validator_test.dart
@@ -1,19 +1,15 @@
-import 'package:dart_foundation/src/validator/bool_validator.dart';
-import 'package:dart_foundation/src/validator/int_validator.dart';
 import 'package:test/test.dart';
 import 'package:dart_foundation/dart_foundation.dart';
-import 'package:dart_foundation/src/validator/string_validator.dart';
-
 void main() {
   group('Validator', () {
-    test('string() should return an instance of StringValidator', () {
-      final result = Validator.string();
-      expect(result, isA<StringValidator>());
+    test('bool() should return an instance of BoolValidator', () {
+      final result = Validator.bool();
+      expect(result, isA<BoolValidator>());
     });
 
-    test('string() should return a new StringValidator instance each time', () {
-      final first = Validator.string();
-      final second = Validator.string();
+    test('bool() should return a new BoolValidator instance each time', () {
+      final first = Validator.bool();
+      final second = Validator.bool();
       expect(identical(first, second), isFalse);
     });
 
@@ -28,14 +24,14 @@ void main() {
       expect(identical(first, second), isFalse);
     });
 
-    test('bool() should return an instance of BoolValidator', () {
-      final result = Validator.bool();
-      expect(result, isA<BoolValidator>());
+    test('string() should return an instance of StringValidator', () {
+      final result = Validator.string();
+      expect(result, isA<StringValidator>());
     });
 
-    test('bool() should return a new BoolValidator instance each time', () {
-      final first = Validator.bool();
-      final second = Validator.bool();
+    test('string() should return a new StringValidator instance each time', () {
+      final first = Validator.string();
+      final second = Validator.string();
       expect(identical(first, second), isFalse);
     });
   });