+++ CHANGELOG.md
@@ -3,6 +3,10 @@ ___
 
 Copyright: Christian Holler
 
+## 1.0.1
+- Added new validators and validation methods.
+- Added DateTime as ValueObject and introduced validated factory.
+
 ## 1.0.0+3
 - Bugfix: Changed order of validation for passwords